requirements.txt
pyspark==3.5.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
delta-spark==3.2.0
